fix.8. Open first position.

+++ src/main/java/ru/grnk/tradevisor/trade/dto/TrvOrder.java
@@ -2,13 +2,15 @@ package ru.grnk.tradevisor.trade.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record TrvOrder(
         String tickerCode,
         int direction,
-        Float activation,
-        float price,
-        float lot,
+        BigDecimal activation,
+        BigDecimal price,
+        int lot,
         boolean isGtc,
         String status
 ) {
